Move organization name to own component
