+++ src/pages/messages/components/MessageList.tsx
@@ -2,7 +2,6 @@ import { Box, Divider, Skeleton, Tooltip, Typography } from '@mui/material';
 import { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
-import { useFetchOrganization } from '../../../api/hooks/useFetchOrganization';
 import { useFetchUser } from '../../../api/hooks/useFetchUser';
 import { ErrorBoundary } from '../../../components/ErrorBoundary';
 import { StyledTruncatableTypography } from '../../../components/styled/Wrappers';
@@ -10,8 +9,8 @@ import { RootState } from '../../../redux/store';
 import { Ticket } from '../../../types/publication_types/ticket.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { toDateString } from '../../../utils/date-helpers';
-import { getLanguageString } from '../../../utils/translation-helpers';
 import { getFullName } from '../../../utils/user-helpers';
+import { MessageItemOrganization } from './MessageItemOrganization';
 import { MessageMenu } from './MessageMenu';
 import { ticketColor } from './TicketListItem';
 
@@ -68,16 +67,21 @@ interface MessageItemProps {
   username: string;
   backgroundColor: string;
   menuElement?: ReactNode;
+  showOrganization?: boolean;
 }
 
-export const MessageItem = ({ text, date, username, backgroundColor, menuElement }: MessageItemProps) => {
+export const MessageItem = ({
+  text,
+  date,
+  username,
+  backgroundColor,
+  menuElement,
+  showOrganization = false,
+}: MessageItemProps) => {
   const { t } = useTranslation();
 
   const senderQuery = useFetchUser(username);
   const senderName = getFullName(senderQuery.data?.givenName, senderQuery.data?.familyName);
-
-  const organizationQuery = useFetchOrganization(senderQuery.data?.institutionCristinId ?? '');
-  const organizationName = getLanguageString(organizationQuery.data?.labels);
 
   return (
     <Box
@@ -108,17 +112,7 @@ export const MessageItem = ({ text, date, username, backgroundColor, menuElement
         </Typography>
         {menuElement}
 
-        <Tooltip title={organizationName}>
-          <StyledTruncatableTypography sx={{ gridColumn: '1/-1' }}>
-            {organizationQuery.isPending ? (
-              <Skeleton sx={{ width: '80%' }} />
-            ) : organizationName ? (
-              organizationName
-            ) : (
-              <i>{t('common.unknown')}</i>
-            )}
-          </StyledTruncatableTypography>
-        </Tooltip>
+        {showOrganization && <MessageItemOrganization organizationId={senderQuery.data?.institutionCristinId ?? ''} />}
       </Box>
 
       <Divider sx={{ mb: '0.5rem', bgcolor: 'primary.main' }} />
